+++ CHANGELOG.md
@@ -24,6 +24,7 @@ added/removed (**Incompatible change**).
 - `get_api` returns `u32` instead `Result<u32>` (**Incompatible change**).
 - `ScmpArch::native()` panics instead of returning an error (**Incompatible change**).
 - `ScmpNotifData.syscall` is now `ScmpSyscall` instead of `i32` (**Incompatible change**).
+- Marked `ScmpNotifData`, `ScmpNotifReq` and `ScmpNotifResp` `#[non_exhaustive]` (**Incompatible change**).
 
 ### Deprecated
 - `NOTIF_FLAG_CONTINUE` use `ScmpNotifRespFlags::CONTINUE.bits()` instead.